Add automatic rerender when user sends message
