+++ src/components/MessageForm/MessageForm.js
@@ -1,16 +1,19 @@
 import React, { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
+import { findMatchingMessages } from '../../utils/messagingAPICalls';
+import { loadMessages } from '../../actions';
 import './MessageForm.scss';
 
-const MessageForm = ({ recipient }) => {
+const MessageForm = ({ otherMessenger }) => {
   const currentUser = useSelector(state => state.currentUser);
+  const dispatch = useDispatch();
   const [messageToSend, setMessageToSend] = useState('');
 
   const sendMessage = (e) => {
     e.preventDefault();
 
     const mutation = {
-      query: "mutation {\n createMessage(input: {\n senderId: \"" + currentUser.id + "\"\n recipientId: \"" + recipient.id + "\"\n body: \"" + messageToSend + "\"\n}) {\n message {\n body \n userId\n conversation {\n recipientId\n }\n }\n }\n }",
+      query: "mutation {\n createMessage(input: {\n senderId: \"" + currentUser.id + "\"\n recipientId: \"" + otherMessenger.id + "\"\n body: \"" + messageToSend + "\"\n}) {\n message {\n body \n userId\n conversation {\n recipientId\n }\n }\n }\n }",
       variables: {}
     };
 
@@ -33,6 +36,17 @@ const MessageForm = ({ recipient }) => {
     .then (data => data.data.createMessage.message)
     .catch(error => console.log(error))
   };
+
+  findMatchingMessages(currentUser.id, otherMessenger.id)
+    .then(data => dispatch(loadMessages({
+      otherMessenger: {
+        id: otherMessenger.id,
+        name: otherMessenger.name,
+        pic: otherMessenger.pic
+      },
+      messages: data.data.messages
+    })))
+    .catch(error => console.log(error))
 
   return (
     <form>
